Replace existing ModalRadioItem with new one
