--- packages/web/src/components/modal-radio/ModalRadioGroup.tsx
@@ -1,9 +1,0 @@
-import { ReactNode } from 'react'
-
-type ModalRadioGroupProps = {
-  items: ReactNode[]
-}
-
-export const ModalRadioGroup = ({ items }: ModalRadioGroupProps) => {
-  return <div>{items.map((item) => item)}</div>
-}
