+++ packages/web/src/components/modal-radio/ModalRadioItem.tsx
@@ -1,35 +1,42 @@
 import { ReactNode } from 'react'
 
+import { RadioButton } from '@audius/stems'
 import cn from 'classnames'
 
-import styles from './ModalRadio.module.css'
+import styles from './ModalRadioItem.module.css'
 
 type ModalRadioItemProps = {
-  children: ReactNode
-  selected: boolean
-  onClick: (e: any) => void
+  label: string
+  title?: ReactNode
+  description: ReactNode
+  value: any
   disabled?: boolean
-  className?: string
+  icon?: ReactNode
+  checkedContent?: ReactNode
 }
 
-export const ModalRadioItem = ({
-  selected,
-  onClick,
-  children,
-  disabled = false,
-  className
-}: ModalRadioItemProps) => {
+export const ModalRadioItem = (props: ModalRadioItemProps) => {
+  const { icon, label, title, description, value, disabled, checkedContent } =
+    props
   return (
-    <div
-      className={cn(
-        styles.radioItem,
-        { [styles.selected]: selected, [styles.disabled]: disabled },
-        className
-      )}
-    >
-      <div className={styles.radioItemContent} onClick={onClick}>
-        {children}
+    <label className={cn(styles.root)}>
+      <RadioButton
+        className={styles.radio}
+        inputClassName={styles.input}
+        aria-label={label}
+        value={value}
+        disabled={disabled}
+      />
+      <div className={styles.labelContent}>
+        <div className={styles.optionTitle}>
+          {icon}
+          <span>{title ?? label}</span>
+        </div>
+        <div className={styles.optionDescription}>{description}</div>
+        {checkedContent ? (
+          <div className={styles.checkedContent}>{checkedContent}</div>
+        ) : null}
       </div>
-    </div>
+    </label>
   )
 }
